% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_iterate.R
\name{di_iterate}
\alias{di_iterate}
\title{Iteratively calculate disproportionate impact via the percentage point gap (PPG) method for many variables.}
\usage{
di_iterate(data, success_vars, group_vars, cohort_vars = NULL,
  scenario_repeat_by_vars = NULL, exclude_scenario_df = NULL,
  weight_var = NULL, ppg_reference_groups = "overall",
  min_moe = 0.03, use_prop_in_moe = FALSE, prop_sub_0 = 0.5,
  prop_sub_1 = 0.5, di_prop_index_cutoff = 0.8,
  di_80_index_cutoff = 0.8, di_80_index_reference_groups = NA)
}
\arguments{
\item{data}{A data frame for which to iterate DI calculation for a set of variables.}

\item{success_vars}{A character vector of success variable names to iterate across.}

\item{group_vars}{A character vector of group (disaggregation) variable names to iterate across.}

\item{cohort_vars}{(Optional) A character vector of the same length as \code{success_vars} to indicate the cohort variable to be used for each variable specified in \code{success_vars}.  A vector of length 1 could be specified, in which case the same cohort variable is used for each success variable.  If not specified, then a single cohort is assumed.}

\item{scenario_repeat_by_vars}{(Optional) A character vector of variables to repeat DI calculations for across all combination of these variables, including '- All' as a group for each variable.  The reference rate used for DI comparison differs for every combination of the variables listed here.}

\item{weight_var}{(Optional) A character scalar specifying the weight variable if the input data set is summarized (ie,  the the success variables specified in `success_vars` contain count of successes).  Weight here corresponds to the denominator when calculating the success rate.  Defaults to `NULL` for an input data set where each row describes each individual.}

\item{ppg_reference_groups}{Either 'overall', 'hpg', 'all but current', or a character vector of the same length as `group_vars` that indicates the reference group value for each group variable in `group_vars` when determining disproportionate impact using the percentage point gap method.}

\item{min_moe}{The minimum margin of error to be used in the PPG calculation, passed to `di_ppg`.}

\item{use_prop_in_moe}{Whether the estimated proportions should be used in the margin of error calculation by the PPG, passed to `di_ppg`.}

\item{prop_sub_0}{Passed to `di_ppg`; defaults to 0.50.}

\item{prop_sub_1}{Passed to `di_ppg`; defaults to 0.50.}

\item{di_prop_index_cutoff}{Threshold used for determining disproportionate impact using the proportionality index; passed to `di_prop_index`; defaults to 0.80.}

\item{di_80_index_cutoff}{Threshold used for determining disproportionate impact using the 80\% index; passed to `di_80_index`; defaults to 0.80.}

\item{di_80_reference_groups}{A character vector of the same length as `group_vars` that indicates the reference group value for each group variable in `group_vars` when determining disproportionate impact using the 80\% index; defaults to \code{NA} (highest performing group as reference).}
}
\value{
A summarized data set (data frame) consisting of: success_variable (elements of `success_vars`), disaggregation (elemeents of `group_vars`), cohort (values corresponding to the variables specified in `cohort_vars`, di_indicator_ppg (1 if there is disproportionate impact per the percentage point gap method, 0 otherwise), di_indicator_prop_index (1 if there is disproportionate impact per the proportionality index, 0 otherwise), di_indicator_80_index (1 if there is disproportionate impact per the 80\% index, 0 otherwise), and other relevant fields returned from `di_ppg`, `di_prop_index`,  and `di_80_index`.
}
\description{
Iteratively calculate disproportionate impact via the percentage point gap (PPG) method for many disaggregation variables.
}
\details{
Iteratively calculate disproportionate impact via the percentage point gap (PPG) method for all combinations of `success_vars`, `group_vars`, and `cohort_vars`, for each combination of subgroups specified by `scenario_repeat_by_vars`.
}
\examples{
library(dplyr)
data(student_equity)
# Multiple group variables
di_iterate(data=student_equity, success_vars=c('Transfer')
  , group_vars=c('Ethnicity', 'Gender'), cohort_vars=c('Cohort')
  , ppg_reference_groups='overall')
}
