% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_point_gap.R
\name{di_ppg}
\alias{di_ppg}
\title{Calculate Disproportionate Impact per the Percentage Point Gap (PPG) method.}
\usage{
di_ppg(success, group, cohort, reference = c("overall", "hpg"), data,
  min.moe = 0.03, use.prop.in.moe = FALSE)
}
\arguments{
\item{success}{A vector of success indicators (\code{1}/\code{0} or \code{TRUE}/\code{FALSE}) or an unquoted reference (name) to a column in \code{data} if it is specified.}

\item{group}{A vector of group names of the same length as \code{success} or an unquoted reference (name) to a column in \code{data} if it is specified.}

\item{cohort}{(Optional) A vector of cohort names of the same length as \code{success} or an unquoted reference (name) to a column in \code{data} if it specified.  Disproportionate Impact is calculated for every group within each cohort.  When \code{cohort} is not specified, then the analysis assumes a single cohort.}

\item{reference}{Either \code{'overall'} (default), \code{'hpg'} (highest performing group), a single proportion (eg, 0.50), or a vector of proportions.  Reference is used as a point of comparison for Disproportionate Impact for each group.  When \code{cohort} is specified:\cr
1. \code{'overall'} will use the overall success rate of each cohort group as the reference;\cr
2. \code{'hpg'} will use the highest performing group in each cohort as reference;\cr
3. the specified proportion will be used for all cohorts;\cr
4. the specified vector of proportions will refer to the reference point for each cohort in alphabetical order (so the number of proportions should equal to the number of unique cohorts).\cr}

\item{data}{(Optional) A data frame containing the variables of interest.  If \code{data} is specified, then \code{success}, \code{group}, and \code{cohort} will be searched within it.}

\item{min.moe}{The minimum margin of error (MOE) to be used in the calculation of Disproportionate Impact and is passed to \link{ppg_moe}.  Defaults to \code{0.03}.}

\item{use.prop.in.moe}{A logical value indicating whether or not the MOE formula should use the observed success rates (\code{TRUE}).  Defaults to \code{FALSE}, which uses 0.50 as the proportion in the MOE formula.  If \code{TRUE}, the success rates are passed to the \code{proportion} argument of \link{ppg_moe}.}
}
\value{
A data frame consisting of: cohort (if used), group, n (sample size), success (number of successes for the cohort-group), pct (proportion of successes for the cohort-group), reference (reference used in DI calculation), moe (margin of error), pct.lo (lower 95\% confidence interval for pct), pct.hi (upper 95\% confidence interval for pct), and di.indicator (1 if there is Disproportionate Impact).
}
\description{
Calculate Disproportionate Impact per the Percentage Point Gap (PPG) method.
}
\details{
This function determines Disproportionate Impact based on the Percentage Point Gap (PPG) method, as described in \href{http://extranet.cccco.edu/Portals/1/TRIS/Research/Analysis/PercentagePointGapMethod2017.pdf}{this} reference from the California Community Colleges Chancellor's Office.  It assumes that a higher rate is good ("success").  For rates that are deemed negative (eg, rate of drop-outs, high is bad), then consider looking at the converse of the non-success (eg, non drop-outs, high is good) instead in order to leverage this function properly.  Note that the margin of error (MOE) is calculated using using 1.96*sqrt(0.25^2/n), with a \code{min.moe} used as the minimum by default.
}
