% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_point_gap.R
\name{ppg_moe}
\alias{ppg_moe}
\title{Margin of Error for the PPG}
\usage{
ppg_moe(n, proportion, min.moe = 0.03)
}
\arguments{
\item{n}{Sample size for the group of interest.}

\item{proportion}{The proportion of successes for the group of interest.  If specified, then the proportion is used in the MOE formula.  Otherwise, a default proportion of 0.50 is used (conservative and yields the maximum MOE).}

\item{min.moe}{The minimum MOE returned even if the sample size is large.  Defaults to 0.03.}
}
\value{
The margin of error for the PPG given the specified sample size.
}
\description{
Calculate the Margin of Error (MOE) for the Percentage Point Gap (PPG) method.
}
