% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_student_equity.R
\docType{data}
\name{student_equity}
\alias{student_equity}
\title{Fake data on student equity}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20000 rows and 4 columns.}
\usage{
data(student_equity)
}
\description{
Data randomly generated to illustrate the use of the package.
}
\examples{
data(student_equity)
}
\keyword{datasets}
